airflow/requirements.txt
certifi
charset-normalizer
docopt
hdfs
idna
numpy
pandas
pyarrow
python-dateutil
pytz
requests
six
sodapy
tzdata
urllib3
hdfs
sqlalchemy
python-dotenv
pyspark
scikit-learn
unidecode
plotly